requirements.txt
blinker
bs4
flask
flask-sqlalchemy
flask-login 
flask-bcrypt
flask-wtf
flask-migrate
flask-script
pytest-flask
sqlalchemy
